requirements.txt
scikit-learn>=1.3.0
shap>=0.44.0
joblib>=1.3.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
httpx>=0.25.0
python-dotenv>=1.0.0
colorama>=0.4.6
flask>=3.0.0
